requirements.txt
lxml
requests
click
grab
guessit
slugify
formic-py3